PageForm: remove slug (only used in one place, could be replaced by pageId)
